+++ kernel/sys/ethernet/ethernet_util.c
@@ -9,14 +9,14 @@
 #include <sys/deviceapi/deviceapi_ethernet.h>
 #include <sys/ethernet/ethernet_util.h>
 
-void ethernet_read(struct device* dev, uint8_t* data, uint32_t size) {
+void ethernet_read(struct device* dev, uint8_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
     ASSERT_NOT_NULL(dev->api);
     struct deviceapi_ethernet* rtc_api = (struct deviceapi_ethernet*)dev->api;
 }
 
-void ethernet_write(struct device* dev, uint8_t* data, uint32_t size) {
+void ethernet_write(struct device* dev, uint8_t* data, uint16_t size) {
     ASSERT_NOT_NULL(dev);
     ASSERT_NOT_NULL(data);
     ASSERT_NOT_NULL(dev->api);
